requirements.txt
fastapi
uvicorn
python-dotenv
aiosqlite
httpx
